requirements.txt
python-telegram-bot[job-queue]>=21.0
google-genai>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
PyGithub>=2.0.0
requests>=2.28.0
